Connection: intercept the handshake
This is the first step towards doing something useful; this way, we
can decide which MySQL server to connect to based on information from
the HandshakeResponse.

This currently forces both peers to authenticate using
"mysql_clear_password"; for the connection to our client, this is
fine, because that connection is supposed to be made over a local
socket (and the password may be ignored anyway because the process is
authenticated via local socket credentials); but for the connection to
the server, we should implement "mysql_native_password".