Connection: schedule EPOLLOUT to retry short send
This fixes a stall bug that can occur when the kernel send buffer becomes
full and the last send() submitted only a portion of the buffer.  In
that case, myproxy would never continue the transfer.  By registering
EPOLLOUT, we get a callback when the kernel is ready for more data.