Connection: adjust sequence_id of forwarded OK/ERR packets during handshake
If the MySQL server has requested an AuthSwitch, then the sequence ids
between client and server and off by two.  Instead of forwarding those
packets as-is, we need to generate new ones with the correct sequence
ids (and with the correct format that may differ due to different
capabilities flags).